# Improve memory allocation behavior by replacing sync.Pool with custom pool implementation